# Approximate FSE encoding costs for selection