… lib/compress/zstd_opt.c