Estimate the cost for using FSE modes `set_basic`, `set_compressed`, and
`set_repeat`, and select the one with the lowest cost.

* The cost of `set_basic` is computed using the cross-entropy cost
  function `ZSTD_crossEntropyCost()`, using the normalized default count
  and the count.
* The cost of `set_repeat` is computed using `FSE_bitCost()`. We check the
  previous table to see if it is able to represent the distribution.
* The cost of `set_compressed` is computed with the entropy cost function
  `ZSTD_entropyCost()`, together with the cost of writing the normalized
  count `ZSTD_NCountCost()`.


# Benchmarks

## Small data at high compression levels

Silesia800: `zstd silesia.tar -B800 -b19e19i15`
GitHub: `zstd -r github/ -b19e19i15`

Silesia800 | Ratio | Compression MB/s | Decompression MB/s
-- | -- | -- | --
Before | 1.897 | 10.6 | 228.2
After | 1.930 | 10.5 | 208.6

GitHub | Ratio | Compression MB/s | Decompression MB/s
-- | -- | -- | --
Before | 3.002 | 10.1 | 265.5
After | 3.060 | 9.70 | 251.2



## Dictionary with well adapted tables

"dict6" and "dict19" are created with the commands:
```
zstd --train-cover=d=8,steps=32 -T2 -r github/  -6 -o dict6
zstd --train-cover=d=8,steps=32 -T2 -r github/  -19 -o dict19
zstd --train-cover=d=8,steps=32 -T2 -r hg-changelog/  -6 -o hg-changelog-dict
zstd --train-cover=d=8,steps=32 -T2 -r hg-commands/  -19 -o hg-commands-dict
zstd --train-cover=d=8,steps=32 -T2 -r hg-manifest/  -19 -o hg-manifest-dict
```
GitHub6: `zstd -r github/ -b6e6i15 -D dict6`
GitHub19: `zstd -r github/ -b19e19i15 -D dict19`
hg-changelog: `zstd -r hg-changelog/ -b19e19i15 -D hg-changelog-dict`
hg-commands: `zstd -r hg-commands/ -b19e19i15 -D hg-commands-dict`
hg-manifest: `zstd -r hg-manifest/ -b19e19i15 -D hg-manifest-dict`


GitHub6 | Ratio | Compression MB/s | Decompression MB/s
-- | -- | -- | --
Before | 10.71 | 2.00 | 4.81.2
After | 10.73 | 1.97 | 476.9


GitHub19 | Size | Ratio | Compression MB/s | Decompression MB/s
-- | -- | -- | -- | --
Before | 667143 | 11.22 | 0.84 | 477.1
After | 667073 | 11.22 | 0.84 | 481.5


hg-changelog | Size | Ratio | Compression MB/s | Decompression MB/s
-- | -- | -- | -- | --
Before | 3787724 | 2.421 | 1.39 | 324.3
After | 3784648 | 2.423 | 1.44 | 316.5


hg-commands | Size | Ratio | Compression MB/s | Decompression MB/s
-- | -- | -- | -- | --
Before | 864908 | 5.241 | 2.53 | 606.7
After | 864675 | 5.242 | 2.53 | 582.9


hg-manifest | Size | Ratio | Compression MB/s | Decompression MB/s
-- | -- | -- | -- | --
Before | 14809800 | 2.884 | 2.92 | 555.2
After | 14802424 | 2.886 | 2.90 | 561.4



## Dictionary with bad tables

"dict" is created with this command using  "dict19" from before. This command basically takes the tables from an unrelated dictionary and replaces the "dict19" tables with them, which means the FSE tables are not adapted well to the GitHub data. The Silesia800 test is run using the GitHub dictionary (or none).
```
head -c 4K silesia.tar > silesia4M
zstd --train -B4K silesia4M -o tmpDict
head -c 500 tmpDict > dict && rm tmpDict
tail -c +501 dict19 >> dict
```

GitHub6: `zstd -r github/ -b6e6i15 -D dict`
GitHub19: `zstd -r github/ -b19e19i15 -D dict`
Silesia800: `zstd -r silesia4M -B800 -b19e19i15 [-D dict19]`


GitHub6 | Ratio | Compression MB/s | Decompression MB/s
-- | -- | -- | --
Before | 8.840 | 1.99 | 487.1
After | 9.714 | 1.96 | 477.6

GitHub19 | Ratio | Compression MB/s | Decompression MB/s
-- | -- | -- | --
Before | 9.109 | 0.90 | 459.7
After | 9.788 | 0.90 | 447.3

Silesia800 | Ratio | Compression MB/s | Decompression MB/s
-- | -- | -- | --
Before | 1.519 | 0.98 | 157.2
After | 1.683 | 0.97 | 132.7
No Dict | 1.685 | 14.3 | 198.2


## Large data

Silesia6: `zstd silesia.tar -b6e6i15`
Silesia19: `zstd silesia.tar -b19e19i15`


Silesia6 | Size | Ratio | Compression MB/s | Decompression MB/s
-- | -- | -- | -- | --
Before | 62397349 | 3.397 | 50.2 | 657.9
After | 62388086 | 3.398 | 52.4 | 656.2

Silesia19 | Size | Ratio | Compression MB/s | Decompression MB/s
-- | -- | -- | -- | --
Before | 53368379 | 3.972 | 1.97 | 550.8
After | 53365440 | 3.972 | 2.12 | 548.9